// library-app/src/app/services/library-api/model/bookCreate.ts
/**
 * Library Web Application
 * A web application for library management.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BookCreate { 
    title: string;
    author: string;
    isbn: string;
    publisher?: string | null;
    edition?: string | null;
    barcode: string;
    total_quantity: number;
}